refactor: split Step1 into GameSidebar and TemplateBrowser components
Extract the 280-line Step1 into three focused files:
- Step1: thin coordinator with context hooks, shared memos, and layout
- GameSidebar: left panel with game search, list, and SidebarItem
- TemplateBrowser: right panel with template search, tag filter, and list
